dom/network/tests/test_networkstats_enabled_no_perm.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test to ensure NetworkStats enabled and no networkstats-manage perm does not allow open</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

// Test to ensure NetworkStats is enabled but mozNetworkStats.getAvailableNetworks
// does not work in content because mozNetworkStats is null when no permission.
  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [['dom.mozNetworkStats.enabled', true]]},
                            function() {
    SpecialPowers.pushPermissions([{ 'type': 'networkstats-manage', 'allow': 0, 'context': window.document }], runTest);
  });

  function runTest() {
    ok(SpecialPowers.getBoolPref("dom.mozNetworkStats.enabled"),
      "Preference 'dom.mozNetworkStats.enabled' is true.");

    ok(!SpecialPowers.hasPermission("networkstats-manage", document),
      "Has no permission 'networkstats-manage'.");

    ok(!('mozNetworkStats' in navigator),
       "navigator.mozNetworkStats should not exist when permission not set");

    SimpleTest.finish();
  }
</script>
</pre>
</body>
</html>
